# Speed-up repeat for AbstractArrays