I'd say this should fix #15553. Compared to the OP in that issue I get times of ~3ms vs ~600ms on master.

Also, compared to this https://github.com/JuliaLang/julia/issues/20495#issuecomment-278041732, I get
```julia
julia> for N in (10, 50, 100)
           A = rand(N, N)
           outer = (N, N)
           @btime repmat($A, $N, $N);
           @btime repeat($A, outer=$outer);
       end
  12.424 μs (202 allocations: 84.45 KiB)
  40.933 μs (130 allocations: 82.39 KiB)
  8.456 ms (5002 allocations: 47.84 MiB)
  12.419 ms (370 allocations: 47.70 MiB)
  169.579 ms (20002 allocations: 763.55 MiB)
  200.986 ms (670 allocations: 762.96 MiB)
```

This `repeat` is still type-unstable, but that should improve when inference and keyword arguments improve.