@nanosoldier 2017-02-17
[Your benchmark job](https://github.com/JuliaLang/julia/pull/20635#issuecomment-280472032) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/2c29ed39007802571ab5bc32c52a516ba8c215f2/b9d83c9_vs_3222749/report.md). cc @jrevels