@nalimilan 2017-02-19
This case broke:
```julia
julia> repeat(Int[], outer=0)
ERROR: BoundsError: attempt to access 0-element UnitRange{Int64} at index [0]
Stacktrace:
 [1] throw_boundserror(::UnitRange{Int64}, ::Int64) at ./abstractarray.jl:417
 [2] getindex at ./range.jl:460 [inlined]
 [3] _repeat at ./abstractarraymath.jl:423 [inlined]
 [4] #repeat#110(::Tuple{Int64}, ::Int64, ::Function, ::Array{Int64,1}) at ./abstractarraymath.jl:365
 [5] (::Base.#kw##repeat)(::Array{Any,1}, ::Base.#repeat, ::Array{Int64,1}) at ./<missing>:0
```

This triggers a failure in DataTables.