@braydenware 2017-07-10
This broke another use case of repeat as well:

```julia
julia> repeat([[1,2], [3,4]]; inner=2)
ERROR: MethodError: Cannot `convert` an object of type Int64 to an object of type Array{Int64,1}
This may have arisen from a call to the constructor Array{Int64,1}(...),
since type constructors fall back to convert methods.
Stacktrace:
 [1] setindex!(::Array{Array{Int64,1},1}, ::Array{Int64,1}, ::UnitRange{Int64}) at ./array.jl:573
 [2] _repeat at ./abstractarraymath.jl:411 [inlined]
 [3] #repeat#117(::Int64, ::Tuple{Int64}, ::Function, ::Array{Array{Int64,1},1}) at ./abstractarraymath.jl:366
 [4] (::Base.#kw##repeat)(::Array{Any,1}, ::Base.#repeat, ::Array{Array{Int64,1},1}) at ./<missing>:0
```
The result should be 
```julia
julia> repeat([[1,2], [3,4]]; inner=2)
4-element Array{Array{Int64,1},1}:
 [1,2]
 [1,2]
 [3,4]
 [3,4]
```
as it was before this commit.

Note that the following works fine though:
```julia
julia> repeat([[1,2],[3,4]]; outer=2)
4-element Array{Array{Int64,1},1}:
 [1, 2]
 [3, 4]
 [1, 2]
 [3, 4]
```