@JeffBezanson 2017-07-10
Thanks @braydenware ; could you file an issue?